+++ plugins/plugin-md-power/src/node/container/fileTree.ts
@@ -1,5 +1,8 @@
+import type { Markdown } from 'vuepress/markdown'
 import Token from 'markdown-it/lib/token.mjs'
+import container from 'markdown-it-container'
 import { removeEndingSlash, removeLeadingSlash } from 'vuepress/shared'
+import { getFileIcon } from '../fileIcons/index.js'
 
 interface FileTreeNode {
   filename: string
@@ -7,6 +10,63 @@ interface FileTreeNode {
   expanded: boolean
   focus: boolean
   empty: boolean
+}
+
+const type = 'file-tree'
+const closeType = `container_${type}_close`
+const componentName = 'FileTreeItem'
+const itemOpen = 'file_tree_item_open'
+const itemClose = 'file_tree_item_close'
+
+export function fileTreePlugin(md: Markdown) {
+  const validate = (info: string): boolean => info.trim().startsWith(type)
+  const render = (tokens: Token[], idx: number): string => {
+    if (tokens[idx].nesting === 1) {
+      const hasRes: number[] = [] // level stack
+      for (
+        let i = idx + 1;
+        !(tokens[i].nesting === -1
+          && tokens[i].type === closeType);
+        ++i
+      ) {
+        const token = tokens[i]
+        if (token.type === 'list_item_open') {
+          const result = resolveTreeNodeInfo(tokens, token, i)
+          if (result) {
+            hasRes.push(token.level)
+            const [info, inline] = result
+            const { filename, type, expanded, empty } = info
+            const icon = getFileIcon(filename, type)
+
+            token.type = itemOpen
+            token.tag = componentName
+            token.attrSet('type', type)
+            token.attrSet(':expanded', expanded ? 'true' : 'false')
+            token.attrSet(':empty', empty ? 'true' : 'false')
+            updateInlineToken(inline, info, icon)
+          }
+          else {
+            hasRes.push(-1)
+          }
+        }
+        else if (token.type === 'list_item_close') {
+          if (token.level === hasRes.pop()) {
+            token.type = itemClose
+            token.tag = componentName
+          }
+        }
+      }
+      const info = tokens[idx].info.trim()
+
+      const title = info.slice(type.length).trim()
+      return `<div class="vp-file-tree">${title ? `<p class="vp-file-tree-title">${title}</p>` : ''}`
+    }
+    else {
+      return '</div>'
+    }
+  }
+
+  md.use(container, type, { validate, render })
 }
 
 export function resolveTreeNodeInfo(
