+++ plugins/plugin-md-power/src/node/fileIcons/definitions.ts
