+++ plugins/plugin-md-power/src/node/fileIcons/index.ts
@@ -1,0 +1,2 @@
+export * from './definitions.js'
+export * from './findIcon.js'
